Second update. The blog's dynamic route problem is not solved yet.
