+++ app/about/layout.tsx
@@ -1,0 +1,8 @@
+import { Metadata } from "next";
+import Page from "./page";  
+
+export const metadata: Metadata = {
+    title: "About Me",
+    description: "This is about me page",
+  };
+export default Page;
