+++ app/blogs/[id]/layout.tsx
@@ -1,0 +1,8 @@
+import { Metadata } from "next";
+import Page from "./page";  
+
+export const metadata: Metadata = {
+    title: "Blog",
+    description: "The blog",
+  };
+export default Page;
