+++ app/blogs/layout.tsx
@@ -1,0 +1,8 @@
+import { Metadata } from "next";
+import Page from "./page";  
+
+export const metadata: Metadata = {
+    title: "Blogs",
+    description: "Lists of blogs",
+  };
+export default Page;
